Removed Python Booleans as default arguments
In order to keep my sanity (Amazon's MWS is not very friendly),
I want to have no booleans or numeric types as defaults because
they will cause an error with urllib.quote(), so in order to avoid
validating all kinds of data, 
I'd rather explicitly define the default argument's value as a string.
Ex. 1 -> "1".
Sorry for the inconvenience.